Code simplification in wallet package.
